refactor(ci): split the Rust toolchain contracts out of ws12_workflow_contracts
Pure move, no behavior change. ws12_workflow_contracts.py had grown to 2,171
lines — past the repo's ~1k ceiling — mixing stress-suite parity, crate scope
filters, WebUI site checks, and this PR's six toolchain validators in one file.
Three separate reviews flagged it (two approach audits, one design lane).

scripts/ci/lib/rust_toolchain_contracts.py now owns one question: what
toolchain, linker, and build flags does a Rust job get? Six validators, their
constants, and a module docstring recording the absence-vs-presence asymmetry
that let a release workflow with no Rust install pass this suite last week.

scripts/ci/lib/workflow_text.py owns the four generic YAML text helpers both
modules need. This absorbed a latent defect: JOB_HEADING was defined TWICE in
ws12 (lines 426 and 988), the second shadowing the first for every validator
below it. The two patterns were `[A-Za-z0-9_-]` and `[a-zA-Z0-9_-]` —
semantically identical, so nothing was wrong today, but only by luck. One
definition now.

The test file imports each symbol from its owning module rather than letting
ws12 act as an implicit re-export, so the ownership is checkable.

classify-test-scope.sh: both new paths join the reborn-scoped list. Without
this the split would have silently changed CI behavior — editing a validator
in its new home would classify differently than editing it in ws12 did — which
is exactly the kind of quiet regression a "pure move" is supposed to not have.
test-classify-test-scope.sh pins it; proven red-first by reverting the list
entry and watching the new assertion fail.

Verified after the move, both validators proven to still fire from their new
home by sabotage: breaking the RUSTUP_TOOLCHAIN export fails the gate, and
deleting the release workflow's composite step fails it with the
cargo-not-found rationale. ws12 self-tests 128 OK, live gate passed,
classify-test-scope 69 PASS / 0 FAIL, check-guidance OK.

ws12_workflow_contracts.py: 2171 -> 1833 lines.
